remove allocations from signal handler
even though GPA might already have pages allocated for me to use, there
still is a possibility it doesn't, and so, it'll malloc() in the signal
handler as I add a SignalData to the maybe_signal_queue

the GPA, by default, has thread-safety which implies the use of a lock.

If we get a signal midway through the allocation, the signal handler
WILL deadlock. we do not want that

thankfully, write() is async-signal-safe, which means we can use it to
deliver the full signal info